Moved cell styling
