+++ src/components/styled/Wrappers.tsx
@@ -1,14 +1,4 @@
-import {
-  Box,
-  BoxProps,
-  FormGroup,
-  ListItem,
-  ListItemProps,
-  styled,
-  TableCell,
-  Typography,
-  TypographyProps,
-} from '@mui/material';
+import { Box, BoxProps, FormGroup, ListItem, ListItemProps, styled, Typography, TypographyProps } from '@mui/material';
 
 export const StyledRightAlignedFooter = styled(Box)({
   display: 'flex',
@@ -179,7 +169,3 @@ export const StyledNviStatusBox = styled(Box)(({ theme }) => ({
   backgroundColor: theme.palette.background.paper,
   marginBottom: '0.5rem',
 }));
-
-export const CenteredTableCell = styled(TableCell)({
-  textAlign: 'center',
-});
